--- anna16/__init__.py
@@ -1,2 +1,0 @@
-from CopyNumberPredictor import CopyNumberPredictor
-from Preprocessing import Preprocessing
